chore: Rename context to pathContext
