+++ src/SubMenu/index.tsx
@@ -23,7 +23,7 @@ import {
   PathUserContext,
   useKeyPath,
   useMeasure,
-} from '../context/MeasureContext';
+} from '../context/PathContext';
 import { useMenuId } from '../context/IdContext';
 
 export interface SubMenuProps {
